// frontend/src/types/word-row-master.type.ts
export interface WordRowMaster {
  id: number;
  row_word_id?: string;
  id_sen?: string;
  word?: string;
  lemma?: string;
  links?: string;
  morph?: string;
  pos?: string;
  phrase?: string;
  grm?: string;
  ner?: string;
  semantic?: string;
  lang_code?: string;
  create_by?: number;
  approval_by?: number;
  created_at?: string;
  updated_at?: string;
}